fix: preserve host-kit import locality
