--- packages/host-kit/src/internal/durable-file.ts
@@ -1,62 +1,0 @@
-import fs from 'node:fs';
-import path from 'node:path';
-import { lstatIfPresent } from './verified-file.ts';
-import { syncDirectoryBestEffort, withAtomicPublishTempPathSync } from './atomic-file.ts';
-
-export type DurableFilePublishMode = 'replace' | 'link-exclusive';
-
-/** Publishes complete UTF-8 contents with a durable file and directory fence. */
-export function publishDurableFileSync(options: {
-  destination: string;
-  contents: string;
-  mode?: number;
-  publish?: DurableFilePublishMode;
-}): void {
-  const directory = path.dirname(options.destination);
-  withAtomicPublishTempPathSync(options.destination, (temporaryPath) => {
-    let descriptor: number | undefined;
-    let failed = false;
-    let primaryError: unknown;
-    try {
-      assertSafeDestination(options.destination);
-      descriptor = fs.openSync(temporaryPath, 'wx', options.mode ?? 0o600);
-      fs.writeFileSync(descriptor, options.contents, 'utf8');
-      fs.fsyncSync(descriptor);
-      fs.closeSync(descriptor);
-      descriptor = undefined;
-      assertSafeDestination(options.destination);
-      if (options.publish === 'link-exclusive') {
-        fs.linkSync(temporaryPath, options.destination);
-      } else {
-        fs.renameSync(temporaryPath, options.destination);
-      }
-      syncDirectoryBestEffort(directory);
-    } catch (error) {
-      failed = true;
-      primaryError = error;
-    }
-    if (descriptor !== undefined) {
-      try {
-        fs.closeSync(descriptor);
-      } catch (error) {
-        if (!failed) {
-          failed = true;
-          primaryError = error;
-        }
-      }
-    }
-    if (failed) throw primaryError;
-  });
-}
-
-function assertSafeDestination(destination: string): void {
-  const stats = lstatIfPresent(destination);
-  if (stats?.isSymbolicLink()) {
-    throw new Error(`Refusing to replace a durable file symbolic link: ${destination}`);
-  }
-  if (stats && !stats.isFile()) {
-    throw new Error(
-      `Refusing to replace a durable path that is not a regular file: ${destination}`,
-    );
-  }
-}
